LidarSim: completed removal of IfW from LidarSim module
The framework does not permit calling IfW from within the LidarSim module.  So it was removed.

The main reason IfW had originally been added into the the LidarSim module was to allow time shifting the uniform wind.  This is now handled by an extension within the IfW module and all wind position requests are handled through a mesh (also permits VTK visualizations).